+++ ostap/fitting/fithelpers.py
@@ -46,7 +46,7 @@ from   ostap.fitting.utils     import make_name, numcpu, ncpu, get_i
 from   ostap.fitting.roocmdarg import check_arg 
 from   ostap.math.random_ext   import ve_gauss, poisson
 from   ostap.logger.pretty     import pretty_ve
-import ROOT, sys, random
+import ROOT, sys, random, math
 # =============================================================================
 from   ostap.logger.logger   import getLogger
 if '__main__' ==  __name__ : logger = getLogger ( 'ostap.fitting.fithelpers')
@@ -2244,20 +2244,19 @@ class FitHelper(VarMaker) :
         >>> n =  pdf.gen_sample ( 10            ) ## get poissonian 
         >>> n =  pdf.gen_sample ( VE ( 10 , 3 ) ) ## get gaussian stuff
         """
-        if   isinstance ( nEvents , integer_types ) and 1 <= nEvents and not sample : return nEvents         
-        elif isinstance ( nEvents , num_types     ) and 0 <  nEvents :
+        if   isinstance ( nEvents , num_types ) and 0 < nEvents and not sample : return round ( nEvents ) 
+        elif isinstance ( nEvents , num_types ) and 0 < nEvents :
             nn = -1 
-            while nn <= 0 : nn = poisson ( nEvents )  
+            while nn <= 0 : nn = poisson ( nEvents )
             return nn 
         elif isinstance ( nEvents , VE ) and 0 < nEvents.cov2 () and 0 < nEvents.value() + 3 * nEvents.error () : 
             nn = -1  
             while nn <= 0 :
                 mu = ve_gauss ( nEvents )
                 if 0 < mu : nn = poisson ( mu ) if sample else int ( mu ) 
-            return nn 
+            return int ( nn ) 
 
         self.error ( "Can't generate positive number from %s/%s" % ( nEvents , type ( nEvents ) ) )
-        return 
 
     # =========================================================================
     ## get the proper min/max range for the variable