+++ source/include/Ostap/Math.h
@@ -814,12 +814,12 @@ namespace Ostap
     /** round to nearest integer, rounds half integers to nearest even integer 
      *  @author Vanya BELYAEV Ivan.Belyaev
      */
-    long round ( const double x ) ;
+    long round ( const double      x ) ;
     // ========================================================================
     /** round to nearest integer, rounds half integers to nearest even integer 
      *  @author Vanya BELYAEV Ivan.Belyaev
      */
-    inline long round ( const long double x ) { return round ( double ( x ) ) ; }
+    long round ( const long double x ) ; 
     // ========================================================================
     /** round to nearest integer, rounds half integers to nearest even integer 
      *  @author Vanya BELYAEV Ivan.Belyaev